+++ gazeplay/src/main/java/net/gazeplay/DefaultGamesLocator.java
@@ -1,10 +1,11 @@
 package net.gazeplay;
 
-import blocs.Blocs;
 import bubbles.Bubble;
 import creampie.CreamPie;
 import lombok.extern.slf4j.Slf4j;
 import magiccards.Card;
+import net.gazeplay.games.blocs.Blocs;
+import net.gazeplay.games.blocs.BlocsGamesStats;
 import ninja.Ninja;
 import utils.games.stats.*;
